PRIME-2346 Update HA site registration step from 8 steps to 5 steps (#2227)
* Update HA site registration step from 8 steps to 5 steps

* fix unit tests

* bug fix

* bug fix

the address fields now will appear and be marked as required when trying to advance without address.

* fix the discard saving issue

* fix PR review issue

* fix operation hours save even using back button and discard modal not showing up

* fix time display issue

* fix typo

* fix label and update title
