+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/administrator-page/administrator-form-state.class.ts
@@ -31,7 +31,8 @@ export class AdministratorFormState extends AbstractFormState<AdministratorForm>
 
   public buildForm(): void {
     this.formInstance = this.fb.group({
-      healthAuthorityPharmanetAdministratorId: [null, [Validators.required]]
+      healthAuthorityPharmanetAdministratorId: [null, [Validators.required]],
+      healthAuthorityTechnicalSupportId: [null, [Validators.required]]
     });
   }
 }
