+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/administrator-page/administrator-form.model.ts
@@ -1,3 +1,4 @@
 export interface AdministratorForm {
   healthAuthorityPharmanetAdministratorId: number;
+  healthAuthorityTechnicalSupportId: number;
 }
