+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/administrator-page/administrator-page.component.html
@@ -16,6 +16,19 @@
                       optionKey="id"
                       optionLabel="fullName"
                       [availableOptions]="pharmanetAdministrators"></app-options-form>
+    <app-page-subheader2>
+      <ng-container appPageSubheader2Title>Technical Support</ng-container>
+      <ng-container appPageSubheader2Summary>
+        Choose the Technical Support for this site.
+      </ng-container>
+    </app-page-subheader2>
+
+    <app-options-form [form]="formState.form"
+                      fieldLabel="Technical Support"
+                      controlName="healthAuthorityTechnicalSupportId"
+                      optionKey="id"
+                      optionLabel="fullName"
+                      [availableOptions]="technicalSupports"></app-options-form>
   </app-page-section>
 
   <app-page-footer [isInitialEnrolment]="!isCompleted"
