+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/health-auth-care-type-page/health-auth-care-type-form-state.class.ts
@@ -1,6 +1,7 @@
 import { FormBuilder, FormControl, Validators } from '@angular/forms';
 
 import { AbstractFormState } from '@lib/classes/abstract-form-state.class';
+import { FormControlValidators } from '@lib/validators/form-control.validators';
 import { HealthAuthorityService } from '@health-auth/shared/services/health-authority.service';
 import { HealthAuthCareTypeForm } from './health-auth-care-type-form.model';
 
@@ -14,6 +15,10 @@ export class HealthAuthCareTypeFormState extends AbstractFormState<HealthAuthCar
     this.buildForm();
   }
 
+  public get healthAuthorityVendorId(): FormControl {
+    return this.formInstance.get('healthAuthorityVendorId') as FormControl;
+  }
+
   public get healthAuthorityCareTypeId(): FormControl {
     return this.formInstance.get('healthAuthorityCareTypeId') as FormControl;
   }
@@ -23,28 +28,39 @@ export class HealthAuthCareTypeFormState extends AbstractFormState<HealthAuthCar
       return;
     }
 
-    const { healthAuthorityCareTypeId } = this.formInstance.getRawValue();
+    const { healthAuthorityCareTypeId, healthAuthorityVendorId } = this.formInstance.getRawValue();
     const healthAuthorityCareType = this.healthAuthorityService.healthAuthority.careTypes
       .find(hact => hact.id === healthAuthorityCareTypeId);
+    const healthAuthorityVendor = this.healthAuthorityService.healthAuthority.vendors
+      .find(hav => hav.id === healthAuthorityVendorId);
 
-    return { healthAuthorityCareType };
+    return { healthAuthorityCareType, healthAuthorityVendor };
   }
 
   public patchValue(model: HealthAuthCareTypeForm): void {
+    if (!model) {
+      return;
+    }
+
     const healthAuthorityCareTypeId = model.healthAuthorityCareType?.id;
+    const healthAuthorityVendorId = model.healthAuthorityVendor?.id;
     if (!this.formInstance || !healthAuthorityCareTypeId) {
       return;
     }
 
-    this.formInstance.patchValue({ healthAuthorityCareTypeId });
+    this.formInstance.patchValue({ healthAuthorityCareTypeId, healthAuthorityVendorId });
   }
 
   public buildForm(): void {
     this.formInstance = this.fb.group({
       healthAuthorityCareTypeId: [
-        0,
+        null,
         [Validators.required]
-      ]
+      ],
+      healthAuthorityVendorId: [
+        0,
+        [FormControlValidators.requiredIndex]
+      ],
     });
   }
 }
