+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/health-auth-care-type-page/health-auth-care-type-form.model.ts
@@ -1,3 +1,3 @@
 import { HealthAuthoritySite } from '@health-auth/shared/models/health-authority-site.model';
 
-export interface HealthAuthCareTypeForm extends Pick<HealthAuthoritySite, 'healthAuthorityCareType'> {}
+export interface HealthAuthCareTypeForm extends Pick<HealthAuthoritySite, 'healthAuthorityCareType' | 'healthAuthorityVendor'> { }
