+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/health-auth-care-type-page/health-auth-care-type-page.component.html
@@ -35,6 +35,32 @@
       </div>
     </section>
 
+    <section class="mb-3">
+      <app-page-subheader2>
+        <ng-container appPageSubheader2Title>Vendor</ng-container>
+        <ng-container appPageSubheader2Summary>
+          What PharmaNet software vendor does this site use?
+        </ng-container>
+      </app-page-subheader2>
+
+      <mat-radio-group class="radio-group"
+                       formControlName="healthAuthorityVendorId">
+        <mat-radio-button *ngFor="let vendor of vendors"
+                          class="radio-button"
+                          [value]="vendor.id">
+          {{ vendor.vendorCode | configCode: 'vendors' }}
+        </mat-radio-button>
+      </mat-radio-group>
+
+      <app-alert *ngIf="hasNoVendorError"
+                 type="danger"
+                 icon="warning">
+        <ng-container #alertContent
+                      class="alert-content">
+          No vendor has been chosen.
+        </ng-container>
+      </app-alert>
+    </section>
   </form>
 
   <app-page-footer [isInitialEnrolment]="!isCompleted"
