+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/health-auth-care-type-page/health-auth-care-type-page.component.scss
@@ -1,0 +1,9 @@
+.radio-group {
+  display: flex;
+  flex-direction: column;
+  margin: 1rem 0;
+}
+
+.radio-button {
+  margin: .25rem 0;
+}
