--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/site-address-page/site-address-form-state.class.ts
@@ -1,47 +1,0 @@
-import { FormBuilder, FormGroup } from '@angular/forms';
-
-import { AbstractFormState } from '@lib/classes/abstract-form-state.class';
-import { FormUtilsService } from '@core/services/form-utils.service';
-import { SiteAddressForm } from './site-address-form.model';
-
-export class SiteAddressFormState extends AbstractFormState<SiteAddressForm> {
-  public constructor(
-    private fb: FormBuilder,
-    private formUtilsService: FormUtilsService
-  ) {
-    super();
-
-    this.buildForm();
-  }
-
-  public get physicalAddress(): FormGroup {
-    return this.formInstance.get('physicalAddress') as FormGroup;
-  }
-
-  public get json(): SiteAddressForm {
-    if (!this.formInstance) {
-      return;
-    }
-
-    return this.formInstance.getRawValue();
-  }
-
-  public patchValue(model: SiteAddressForm): void {
-    if (!this.formInstance || !model?.physicalAddress) {
-      return;
-    }
-
-    this.formInstance.patchValue(model);
-  }
-
-  public buildForm(): void {
-    this.formInstance = this.fb.group({
-      physicalAddress: this.formUtilsService.buildAddressForm({
-        areRequired: ['street', 'city', 'provinceCode', 'countryCode', 'postal'],
-        areDisabled: ['provinceCode', 'countryCode'],
-        useDefaults: ['provinceCode', 'countryCode'],
-        exclude: ['street2']
-      })
-    });
-  }
-}
