--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/site-address-page/site-address-form.model.ts
@@ -1,5 +1,0 @@
-import { Address } from '@lib/models/address.model';
-
-export interface SiteAddressForm {
-  physicalAddress: Address;
-}
