--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/site-address-page/site-address-overview.component.spec.ts
@@ -1,32 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { RouterTestingModule } from '@angular/router/testing';
-
-import { SiteAddressOverviewComponent } from './site-address-overview.component';
-
-describe('SiteAddressOverviewComponent', () => {
-  let component: SiteAddressOverviewComponent;
-  let fixture: ComponentFixture<SiteAddressOverviewComponent>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule
-      ],
-      declarations: [
-        SiteAddressOverviewComponent
-      ],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA]
-    }).compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(SiteAddressOverviewComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
