--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/site-address-page/site-address-page.component.html
@@ -1,34 +1,0 @@
-<app-page [busy]="busy">
-  <app-page-header>PharmaNet Site Registration</app-page-header>
-  <app-site-progress-indicator [inProgress]="!isCompleted"></app-site-progress-indicator>
-
-  <form (ngSubmit)="onSubmit()"
-        [formGroup]="formState.form"
-        novalidate
-        autocomplete="off">
-
-    <section class="mb-3">
-      <app-page-subheader2>
-        <ng-container appPageSubheader2Title>{{ title }}</ng-container>
-        <ng-container appPageSubheader2Summary>
-          Provide your site’s street or physical address. Do not use a PO Box.
-        </ng-container>
-      </app-page-subheader2>
-
-      <app-address-form [form]="formState.physicalAddress"
-                        [formControlNames]="formControlNames"
-                        [inBc]="true"
-                        [showManualButton]="true"
-                        [showAddressFields]="showAddressFields"></app-address-form>
-    </section>
-
-  </form>
-
-  <app-page-footer [isInitialEnrolment]="!isCompleted"
-                   [hasSecondaryAction]="true"
-                   primaryActionLabel="Save and Continue"
-                   (save)="onSubmit()"
-                   (back)="onBack()"
-                   (continue)="onBack()"></app-page-footer>
-
-</app-page>
