--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/site-address-page/site-address-page.component.spec.ts
@@ -1,78 +1,0 @@
-import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
-import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { RouterTestingModule } from '@angular/router/testing';
-import { ReactiveFormsModule } from '@angular/forms';
-import { ActivatedRoute } from '@angular/router';
-
-import { MockConfigService } from 'test/mocks/mock-config.service';
-import { MockHealthAuthoritySiteService } from 'test/mocks/mock-health-authority-site.service';
-
-import { APP_CONFIG, APP_DI_CONFIG } from 'app/app-config.module';
-import { NgxMaterialModule } from '@lib/modules/ngx-material/ngx-material.module';
-import { ConfigService } from '@config/config.service';
-import { CapitalizePipe } from '@shared/pipes/capitalize.pipe';
-import { HealthAuthoritySiteService } from '@health-auth/shared/services/health-authority-site.service';
-import { SiteAddressPageComponent } from './site-address-page.component';
-
-describe('SiteAddressPageComponent', () => {
-  let component: SiteAddressPageComponent;
-  let fixture: ComponentFixture<SiteAddressPageComponent>;
-  const mockActivatedRoute = {
-    snapshot: {
-      data: {
-        title: 'Site Address',
-      },
-      params: {
-        haid: 1,
-        sid: 7
-      }
-    }
-  };
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [
-        SiteAddressPageComponent
-      ],
-      imports: [
-        BrowserAnimationsModule,
-        HttpClientTestingModule,
-        RouterTestingModule,
-        ReactiveFormsModule,
-        NgxMaterialModule
-      ],
-      providers: [
-        {
-          provide: APP_CONFIG,
-          useValue: APP_DI_CONFIG
-        },
-        {
-          provide: ConfigService,
-          useClass: MockConfigService
-        },
-        {
-          provide: ActivatedRoute,
-          useValue: mockActivatedRoute
-        },
-        {
-          provide: HealthAuthoritySiteService,
-          useClass: MockHealthAuthoritySiteService
-        },
-        CapitalizePipe
-      ],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA]
-    }).compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(SiteAddressPageComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
