--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/site-address-page/site-address-page.component.ts
@@ -1,99 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
-import { MatDialog } from '@angular/material/dialog';
-import { FormBuilder } from '@angular/forms';
-
-import { RouteUtils } from '@lib/utils/route-utils.class';
-import { Address, AddressLine } from '@lib/models/address.model';
-import { FormUtilsService } from '@core/services/form-utils.service';
-import { HealthAuthoritySiteResource } from '@core/resources/health-authority-site-resource.service';
-
-import { HealthAuthSiteRegRoutes } from '@health-auth/health-auth-site-reg.routes';
-import { HealthAuthoritySiteService } from '@health-auth/shared/services/health-authority-site.service';
-import { HealthAuthoritySiteFormStateService } from '@health-auth/shared/services/health-authority-site-form-state.service';
-import { AbstractHealthAuthoritySiteRegistrationPage } from '@health-auth/shared/classes/abstract-health-authority-site-registration-page.class';
-import { SiteAddressFormState } from './site-address-form-state.class';
-
-@Component({
-  selector: 'app-site-address-page',
-  templateUrl: './site-address-page.component.html',
-  styleUrls: ['./site-address-page.component.scss']
-})
-export class SiteAddressPageComponent extends AbstractHealthAuthoritySiteRegistrationPage implements OnInit {
-  public formState: SiteAddressFormState;
-  public title: string;
-  public routeUtils: RouteUtils;
-  public formControlNames: AddressLine[];
-  public isCompleted: boolean;
-  public showAddressFields: boolean;
-  public SiteRoutes = HealthAuthSiteRegRoutes;
-
-  constructor(
-    protected dialog: MatDialog,
-    protected formUtilsService: FormUtilsService,
-    protected route: ActivatedRoute,
-    protected healthAuthoritySiteService: HealthAuthoritySiteService,
-    protected healthAuthoritySiteFormStateService: HealthAuthoritySiteFormStateService,
-    protected healthAuthoritySiteResource: HealthAuthoritySiteResource,
-    private fb: FormBuilder,
-    router: Router
-  ) {
-    super(dialog, formUtilsService, route, healthAuthoritySiteService, healthAuthoritySiteFormStateService, healthAuthoritySiteResource);
-
-    this.title = route.snapshot.data.title;
-    this.routeUtils = new RouteUtils(route, router, HealthAuthSiteRegRoutes.MODULE_PATH);
-
-    this.formControlNames = [
-      'street',
-      'city',
-      'provinceCode',
-      'postal'
-    ];
-  }
-
-  public onBack(): void {
-    const backRoutePath = (this.isCompleted)
-      ? HealthAuthSiteRegRoutes.SITE_OVERVIEW
-      : HealthAuthSiteRegRoutes.SITE_INFORMATION;
-
-    this.routeUtils.routeRelativeTo(backRoutePath);
-  }
-
-  public ngOnInit(): void {
-    this.createFormInstance();
-    this.patchForm();
-  }
-
-  protected createFormInstance(): void {
-    this.formState = this.healthAuthoritySiteFormStateService.siteAddressFormState;
-  }
-
-  protected patchForm(): void {
-    const healthAuthId = +this.route.snapshot.params.haid;
-    const healthAuthSiteId = +this.route.snapshot.params.sid;
-    if (!healthAuthId || !healthAuthSiteId) {
-      throw new Error('No health authority site ID was provided');
-    }
-
-    const site = this.healthAuthoritySiteService.site;
-
-    if (Address.isNotEmpty(site.physicalAddress, ['countryCode', 'provinceCode'])) {
-      this.showAddressFields = true;
-    }
-
-    this.isCompleted = site?.completed;
-    this.healthAuthoritySiteFormStateService.setForm(site, !this.hasBeenSubmitted);
-  }
-
-  protected onSubmitFormIsInvalid(): void {
-    this.showAddressFields = true;
-  }
-
-  protected afterSubmitIsSuccessful(): void {
-    const nextRoutePath = (this.isCompleted)
-      ? HealthAuthSiteRegRoutes.SITE_OVERVIEW
-      : HealthAuthSiteRegRoutes.HOURS_OPERATION;
-
-    this.routeUtils.routeRelativeTo(nextRoutePath);
-  }
-}
