+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/site-information-page/site-information-form.model.ts
@@ -1,3 +1,4 @@
 import { HealthAuthoritySite } from '@health-auth/shared/models/health-authority-site.model';
+import { Address } from '@lib/models/address.model';
 
-export interface SiteInformationForm extends Pick<HealthAuthoritySite, 'siteName' | 'pec' | 'securityGroupCode'> { }
+export interface SiteInformationForm extends Pick<HealthAuthoritySite, 'siteName' | 'pec' | 'securityGroupCode' | 'physicalAddress'> { }
