+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/site-information-page/site-information-page.component.html
@@ -37,8 +37,22 @@
         </div>
       </div>
 
-    </section>
+      <div class="row">
+        <div class="col">
+          <app-page-subheader2>
+            <ng-container appPageSubheader2Summary>
+              Provide your site’s street or physical address. Do not use a PO Box.
+            </ng-container>
+          </app-page-subheader2>
 
+          <app-address-form [form]="formState.physicalAddress"
+                            [formControlNames]="formControlNames"
+                            [inBc]="true"
+                            [showManualButton]="true"
+                            [showAddressFields]="showAddressFields"></app-address-form>
+        </div>
+      </div>
+    </section>
   </form>
 
   <app-page-footer [isInitialEnrolment]="!isCompleted"
