+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/site-information-page/site-information-page.component.spec.ts
@@ -4,12 +4,14 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
+import { ActivatedRoute } from '@angular/router';
 
 import { KeycloakService } from 'keycloak-angular';
 
 import { MockAuthService } from 'test/mocks/mock-auth.service';
 import { MockConfigService } from 'test/mocks/mock-config.service';
 import { MockPermissionService } from 'test/mocks/mock-permission.service';
+import { MockHealthAuthoritySiteService } from 'test/mocks/mock-health-authority-site.service';
 
 import { NgxMaterialModule } from '@lib/modules/ngx-material/ngx-material.module';
 
@@ -21,11 +23,23 @@ import { SiteInformationPageComponent } from './site-information-page.component'
 import { CapitalizePipe } from '@shared/pipes/capitalize.pipe';
 
 import { HealthAuthSiteRegRoutes } from '@health-auth/health-auth-site-reg.routes';
+import { HealthAuthoritySiteService } from '@health-auth/shared/services/health-authority-site.service';
 
 describe('SiteInformationPageComponent', () => {
   let component: SiteInformationPageComponent;
   let fixture: ComponentFixture<SiteInformationPageComponent>;
   let spyOnRouteRelativeTo;
+  const mockActivatedRoute = {
+    snapshot: {
+      data: {
+        title: 'Site Address',
+      },
+      params: {
+        haid: 1,
+        sid: 7
+      }
+    }
+  };
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
@@ -53,6 +67,14 @@ describe('SiteInformationPageComponent', () => {
         {
           provide: PermissionService,
           useClass: MockPermissionService
+        },
+        {
+          provide: ActivatedRoute,
+          useValue: mockActivatedRoute
+        },
+        {
+          provide: HealthAuthoritySiteService,
+          useClass: MockHealthAuthoritySiteService
         },
         KeycloakService,
         CapitalizePipe
@@ -82,11 +104,11 @@ describe('SiteInformationPageComponent', () => {
   });
 
   describe('testing onBack() with an incomplete profile', () => {
-    it('should call routeToRelativePath with HealthAuthSiteRegRoutes.VENDOR', () => {
+    it('should call routeToRelativePath with HealthAuthSiteRegRoutes.HEALTH_AUTH_CARE_TYPE', () => {
       component.isCompleted = false;
 
       component.onBack();
-      expect(spyOnRouteRelativeTo).toHaveBeenCalledWith(HealthAuthSiteRegRoutes.VENDOR);
+      expect(spyOnRouteRelativeTo).toHaveBeenCalledWith(HealthAuthSiteRegRoutes.HEALTH_AUTH_CARE_TYPE);
     });
   });
 });
