+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/site-information-page/site-information-page.component.ts
@@ -4,6 +4,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { FormBuilder } from '@angular/forms';
 
 import { RouteUtils } from '@lib/utils/route-utils.class';
+import { Address, AddressLine } from '@lib/models/address.model';
 import { Config } from '@config/config.model';
 import { ConfigService } from '@config/config.service';
 import { FormUtilsService } from '@core/services/form-utils.service';
@@ -25,7 +26,9 @@ export class SiteInformationPageComponent extends AbstractHealthAuthoritySiteReg
   public title: string;
   public routeUtils: RouteUtils;
   public securityGroups: Config<number>[];
+  public formControlNames: AddressLine[];
   public isCompleted: boolean;
+  public showAddressFields: boolean;
 
   constructor(
     protected dialog: MatDialog,
@@ -44,12 +47,18 @@ export class SiteInformationPageComponent extends AbstractHealthAuthoritySiteReg
     this.routeUtils = new RouteUtils(route, router, HealthAuthSiteRegRoutes.MODULE_PATH);
 
     this.securityGroups = this.configService.securityGroups;
+    this.formControlNames = [
+      'street',
+      'city',
+      'provinceCode',
+      'postal'
+    ];
   }
 
   public onBack(): void {
     const backRoutePath = (this.isCompleted)
       ? HealthAuthSiteRegRoutes.SITE_OVERVIEW
-      : HealthAuthSiteRegRoutes.VENDOR;
+      : HealthAuthSiteRegRoutes.HEALTH_AUTH_CARE_TYPE;
 
     this.routeUtils.routeRelativeTo(backRoutePath);
   }
@@ -67,10 +76,13 @@ export class SiteInformationPageComponent extends AbstractHealthAuthoritySiteReg
     const healthAuthId = +this.route.snapshot.params.haid;
     const healthAuthSiteId = +this.route.snapshot.params.sid;
     if (!healthAuthId || !healthAuthSiteId) {
-      return;
+      throw new Error('No health authority site ID was provided');
     }
 
     const site = this.healthAuthoritySiteService.site;
+    if (Address.isNotEmpty(site.physicalAddress, ['countryCode', 'provinceCode'])) {
+      this.showAddressFields = true;
+    }
     this.isCompleted = site?.completed;
     this.healthAuthoritySiteFormStateService.setForm(site, !this.hasBeenSubmitted);
   }
@@ -78,8 +90,21 @@ export class SiteInformationPageComponent extends AbstractHealthAuthoritySiteReg
   protected afterSubmitIsSuccessful(): void {
     const nextRoutePath = (this.isCompleted)
       ? HealthAuthSiteRegRoutes.SITE_OVERVIEW
-      : HealthAuthSiteRegRoutes.SITE_ADDRESS;
+      : HealthAuthSiteRegRoutes.HOURS_OPERATION;
 
     this.routeUtils.routeRelativeTo(nextRoutePath);
+  }
+
+  protected handleDeactivation(result: boolean): void {
+    if (!result) {
+      return;
+    }
+
+    // Replace previous values on deactivation so updates are discarded
+    this.healthAuthoritySiteFormStateService.patchSiteInformationForm(this.healthAuthoritySiteService.site);
+  }
+
+  protected onSubmitFormIsInvalid(): void {
+    this.showAddressFields = true;
   }
 }
