--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/technical-support-page/technical-support-form.model.ts
@@ -1,3 +1,0 @@
-export interface TechnicalSupportForm {
-  healthAuthorityTechnicalSupportId: number;
-}
