--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/vendor-page/vendor-form.model.ts
@@ -1,3 +1,0 @@
-import { HealthAuthoritySite } from '@health-auth/shared/models/health-authority-site.model';
-
-export interface VendorForm extends Pick<HealthAuthoritySite, 'healthAuthorityVendor'> {}
