--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/vendor-page/vendor-overview.component.spec.ts
@@ -1,47 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { RouterTestingModule } from '@angular/router/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-
-import { APP_CONFIG, APP_DI_CONFIG } from 'app/app-config.module';
-import { ConfigCodePipe } from '@config/config-code.pipe';
-import { DefaultPipe } from '@shared/pipes/default.pipe';
-import { VendorOverviewComponent } from './vendor-overview.component';
-
-describe('VendorOverviewComponent', () => {
-  let component: VendorOverviewComponent;
-  let fixture: ComponentFixture<VendorOverviewComponent>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        HttpClientTestingModule,
-        RouterTestingModule,
-        MatSnackBarModule
-      ],
-      declarations: [
-        VendorOverviewComponent,
-        DefaultPipe,
-        ConfigCodePipe
-      ],
-      providers: [
-        {
-          provide: APP_CONFIG,
-          useValue: APP_DI_CONFIG
-        }
-      ],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA]
-    }).compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(VendorOverviewComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
