--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/vendor-page/vendor-page.component.html
@@ -1,46 +1,0 @@
-<app-page [busy]="busy">
-  <app-page-header>PharmaNet Site Registration</app-page-header>
-  <app-progress-indicator [noContent]="true"></app-progress-indicator>
-
-  <form (ngSubmit)="onSubmit()"
-        [formGroup]="formState.form"
-        novalidate
-        autocomplete="off">
-
-    <section class="mb-4">
-      <app-page-subheader2>
-        <ng-container appPageSubheader2Title>{{ title }}</ng-container>
-        <ng-container appPageSubheader2Summary>
-          What PharmaNet software vendor does this site use?
-        </ng-container>
-      </app-page-subheader2>
-
-      <mat-radio-group class="radio-group"
-                       formControlName="healthAuthorityVendorId">
-        <mat-radio-button *ngFor="let vendor of vendors"
-                          class="radio-button"
-                          [value]="vendor.id">
-          {{ vendor.vendorCode | configCode: 'vendors' }}
-        </mat-radio-button>
-      </mat-radio-group>
-
-      <app-alert *ngIf="hasNoVendorError"
-                 type="danger"
-                 icon="warning">
-        <ng-container #alertContent
-                      class="alert-content">
-          No vendor has been chosen.
-        </ng-container>
-      </app-alert>
-    </section>
-
-  </form>
-
-  <app-page-footer [isInitialEnrolment]="!isCompleted"
-                   [hasSecondaryAction]="true"
-                   primaryActionLabel="Save and Continue"
-                   (save)="onSubmit()"
-                   (back)="onBack()"
-                   (continue)="onBack()"></app-page-footer>
-
-</app-page>
