--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/vendor-page/vendor-page.component.scss
@@ -1,9 +1,0 @@
-.radio-group {
-  display: flex;
-  flex-direction: column;
-  margin: 1rem 0;
-}
-
-.radio-button {
-  margin: .25rem 0;
-}
