--- prime-angular-frontend/src/app/modules/health-auth-site-reg/pages/vendor-page/vendor-page.component.ts
@@ -1,100 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { Location } from '@angular/common';
-import { FormBuilder } from '@angular/forms';
-import { ActivatedRoute, Router } from '@angular/router';
-import { MatDialog } from '@angular/material/dialog';
-
-import { RouteUtils } from '@lib/utils/route-utils.class';
-import { ConfigService } from '@config/config.service';
-import { FormUtilsService } from '@core/services/form-utils.service';
-import { HealthAuthoritySiteResource } from '@core/resources/health-authority-site-resource.service';
-
-import { HealthAuthSiteRegRoutes } from '@health-auth/health-auth-site-reg.routes';
-import { HealthAuthorityVendor } from '@health-auth/shared/models/health-authority-vendor.model';
-import { HealthAuthoritySiteService } from '@health-auth/shared/services/health-authority-site.service';
-import { HealthAuthoritySiteFormStateService } from '@health-auth/shared/services/health-authority-site-form-state.service';
-import { AuthorizedUserService } from '@health-auth/shared/services/authorized-user.service';
-import { AbstractHealthAuthoritySiteRegistrationPage } from '@health-auth/shared/classes/abstract-health-authority-site-registration-page.class';
-import { VendorFormState } from './vendor-form-state.class';
-
-@Component({
-  selector: 'app-vendor-page',
-  templateUrl: './vendor-page.component.html',
-  styleUrls: ['./vendor-page.component.scss']
-})
-export class VendorPageComponent extends AbstractHealthAuthoritySiteRegistrationPage implements OnInit {
-  public formState: VendorFormState;
-  public title: string;
-  public routeUtils: RouteUtils;
-  public vendors: HealthAuthorityVendor[];
-  public isCompleted: boolean;
-  public hasNoVendorError: boolean;
-
-  constructor(
-    protected dialog: MatDialog,
-    protected formUtilsService: FormUtilsService,
-    protected route: ActivatedRoute,
-    protected healthAuthoritySiteService: HealthAuthoritySiteService,
-    protected healthAuthoritySiteFormStateService: HealthAuthoritySiteFormStateService,
-    protected healthAuthoritySiteResource: HealthAuthoritySiteResource,
-    private fb: FormBuilder,
-    private location: Location, // change access modifier to protected?
-    private configService: ConfigService,
-    private authorizedUserService: AuthorizedUserService,
-    router: Router
-  ) {
-    super(dialog, formUtilsService, route, healthAuthoritySiteService, healthAuthoritySiteFormStateService, healthAuthoritySiteResource);
-
-    this.title = this.route.snapshot.data.title;
-    this.routeUtils = new RouteUtils(route, router, HealthAuthSiteRegRoutes.MODULE_PATH, location);
-    this.hasNoVendorError = false;
-  }
-
-  public onBack(): void {
-    const backRoutePath = this.isCompleted
-      ? HealthAuthSiteRegRoutes.SITE_OVERVIEW
-      : HealthAuthSiteRegRoutes.HEALTH_AUTH_CARE_TYPE;
-
-    this.routeUtils.routeRelativeTo(backRoutePath);
-  }
-
-  public ngOnInit(): void {
-    this.createFormInstance();
-    this.patchForm();
-  }
-
-  protected createFormInstance(): void {
-    this.formState = this.healthAuthoritySiteFormStateService.vendorFormState;
-  }
-
-  protected patchForm(): void {
-    const healthAuthId = +this.route.snapshot.params.haid;
-    const healthAuthSiteId = +this.route.snapshot.params.sid;
-    if (!healthAuthId || !healthAuthSiteId) {
-      throw new Error('No health authority site ID was provided');
-    }
-
-    const site = this.healthAuthoritySiteService.site;
-    this.vendors = this.route.snapshot.data.healthAuthority?.vendors ?? [];
-    this.isCompleted = site?.completed;
-    this.healthAuthoritySiteFormStateService.setForm(site, !this.hasBeenSubmitted);
-  }
-
-  protected onSubmitFormIsValid(): void {
-    this.hasNoVendorError = false;
-  }
-
-  protected onSubmitFormIsInvalid(): void {
-    if (!this.formState.healthAuthorityVendorId.value) {
-      this.hasNoVendorError = true;
-    }
-  }
-
-  protected afterSubmitIsSuccessful(): void {
-    const nextRoutePath = (this.isCompleted)
-      ? HealthAuthSiteRegRoutes.SITE_OVERVIEW
-      : HealthAuthSiteRegRoutes.SITE_INFORMATION;
-
-    this.routeUtils.routeRelativeTo(nextRoutePath);
-  }
-}
