+++ prime-angular-frontend/src/app/modules/health-auth-site-reg/shared/models/health-authority-site-create.model.ts
@@ -1,4 +1,5 @@
 export interface HealthAuthoritySiteCreate {
   authorizedUserId: number;
   healthAuthorityCareTypeId: number;
+  healthAuthorityVendorId: number;
 }
