+++ prime-angular-frontend/src/app/shared/components/enrollee/enrollee-review/enrollee-review.component.html
@@ -46,18 +46,21 @@
 <app-address-view title="Verified Address"
                   [address]="enrollee.verifiedAddress"
                   [showRedirect]="showEditRedirect"
+                  [showEditRedirect]="showEditRedirect"
                   [showIfEmpty]="true"
                   (route)="onRoute(demographicRoutePath)"></app-address-view>
 
 <app-address-view title="Physical Address"
                   [address]="enrollee?.physicalAddress"
                   [showRedirect]="showEditRedirect"
+                  [showEditRedirect]="showEditRedirect"
                   [showIfEmpty]="true"
                   (route)="onRoute(demographicRoutePath)"></app-address-view>
 
 <app-address-view title="Mailing Address"
                   [address]="enrollee?.mailingAddress"
                   [showRedirect]="showEditRedirect"
+                  [showEditRedirect]="showEditRedirect"
                   [showIfEmpty]="true"
                   (route)="onRoute(demographicRoutePath)"></app-address-view>
