+++ prime-angular-frontend/src/app/shared/components/site/overview-container/overview-container.component.html
@@ -10,6 +10,7 @@
   <app-address-view title="Physical Address"
                     [address]="organization?.signingAuthority.physicalAddress"
                     [showRedirect]="showEditRedirect"
+                    [showEditRedirect]="showEditRedirect"
                     [showIfEmpty]="true"
                     (route)="onRoute([SiteRoutes.MODULE_PATH, SiteRoutes.ORGANIZATIONS, organization.id, SiteRoutes.ORGANIZATION_SIGNING_AUTHORITY])">
   </app-address-view>
@@ -17,6 +18,7 @@
   <app-address-view title="Mailing Address"
                     [address]="organization?.signingAuthority.mailingAddress"
                     [showRedirect]="showEditRedirect"
+                    [showEditRedirect]="showEditRedirect"
                     [showIfEmpty]="true"
                     (route)="onRoute([SiteRoutes.MODULE_PATH, SiteRoutes.ORGANIZATIONS, organization.id, SiteRoutes.ORGANIZATION_SIGNING_AUTHORITY])">
   </app-address-view>
