+++ prime-dotnet-webapi/ViewModels/Health Authority Sites/HealthAuthoritySiteCreateModel.cs
@@ -3,6 +3,7 @@ namespace Prime.ViewModels.HealthAuthoritySites
     public class HealthAuthoritySiteCreateModel
     {
         public int HealthAuthorityCareTypeId { get; set; }
+        public int healthAuthorityVendorId { get; set; }
         public int AuthorizedUserId { get; set; }
     }
 }
